skeleton/requirements.txt
flask
pytest
gunicorn
suds
spyne
lxml
netifaces
dnspython